update of fast tools and weio

+++ pydatview/fastlib.py
@@ -151,9 +151,21 @@ def run_cmd(input_file, exe, wait=True, ShowOutputs=False, ShowCommand=True):
     return p
 # --- END cmd.py
 
-def run_fastfiles(fastfiles, fastExe=None, parallel=True, ShowOutputs=True, nCores=None, ShowCommand=True):
+def run_fastfiles(fastfiles, fastExe=None, parallel=True, ShowOutputs=True, nCores=None, ShowCommand=True, ReRun=True):
     if fastExe is None:
         fastExe=FAST_EXE
+    if not ReRun:
+        # Figure out which files exist
+        newfiles=[]
+        for f in fastfiles:
+            base=os.path.splitext(f)[0]
+            if os.path.exists(base+'.outb') or os.path.exists(base+'.out'):
+                print('>>> Skipping existing simulation for: ',f)
+                pass
+            else:
+                newfiles.append(f)
+        fastfiles=newfiles
+
     return run_cmds(fastfiles, fastExe, parallel=parallel, ShowOutputs=ShowOutputs, nCores=nCores, ShowCommand=ShowCommand)
 
 def run_fast(input_file, fastExe=None, wait=True, ShowOutputs=False, ShowCommand=True):
@@ -203,6 +215,27 @@ def ED_BldStations(ED):
     r_nodes      = bld_fract*(ED['TipRad']-ED['HubRad']) + ED['HubRad']
     return bld_fract, r_nodes
 
+def ED_TwrStations(ED):
+    """ Returns ElastoDyn Tower Station positions, useful to know where the outputs are.
+    INPUTS:
+       - ED: either:
+           - a filename of a ElastoDyn input file
+           - an instance of FileCl, as returned by reading the file, ED = weio.read(ED_filename)
+
+    OUTPUTS:
+        - r_fract: fraction of the towet length were stations are defined
+        - h_nodes: height from the *ground* of the stations  (not from the Tower base)
+    """
+    if not isinstance(ED,weio.FASTInFile):
+        ED = weio.FASTInFile(ED)
+
+    nTwrNodes = ED['TwrNodes']
+    twr_fract    = np.arange(1./nTwrNodes/2., 1, 1./nTwrNodes)
+    h_nodes      = twr_fract*(ED['TowerHt']-ED['TowerBsHt']) + ED['TowerBsHt']
+    return twr_fract, h_nodes
+
+
+
 def ED_BldGag(ED):
     """ Returns the radial position of ElastoDyn blade gages 
     INPUTS:
@@ -224,6 +257,29 @@ def ED_BldGag(ED):
         Inodes = np.array([ED['BldGagNd']])
     r_gag = r_nodes[ Inodes[:nOuts] -1]
     return r_gag
+
+def ED_TwrGag(ED):
+    """ Returns the heights of ElastoDyn blade gages 
+    INPUTS:
+       - ED: either:
+           - a filename of a ElastoDyn input file
+           - an instance of FileCl, as returned by reading the file, ED = weio.read(ED_filename)
+    OUTPUTS:
+       - h_gag: The heights of the gages, given from the ground height (tower base + TowerBsHt)
+    """
+    if not isinstance(ED,weio.FASTInFile):
+        ED = weio.FASTInFile(ED)
+    _,h_nodes= ED_TwrStations(ED)
+    nOuts = ED['NTwGages']
+    if nOuts<=0:
+        return np.array([])
+    if type(ED['TwrGagNd']) is list:
+        Inodes = np.asarray(ED['TwrGagNd'])
+    else:
+        Inodes = np.array([ED['TwrGagNd']])
+    h_gag = h_nodes[ Inodes[:nOuts] -1]
+    return h_gag
+
 
 def AD14_BldGag(AD):
     """ Returns the radial position of AeroDyn 14 blade gages (based on "print" in column 6)
@@ -538,9 +594,27 @@ def copyTree(src, dst):
                 forceMergeFlatDir(s, d)
 
 
-def templateReplaceGeneral(template_dir, PARAMS, workdir=None, main_file=None, name_function=None, RemoveAllowed=False):
-    """ Replace parameters in a fast folder using a list of dictionaries where the keys are for instance:
-        'FAST|DT', 'EDFile|GBRatio', 'ServoFile|GenEff'
+def templateReplaceGeneral(PARAMS, template_dir=None, output_dir=None, main_file=None, name_function=None, RemoveAllowed=False):
+    """ Generate inputs files by replacing different parameters from a template file.
+    The generated files are placed in the output directory `output_dir` 
+    The files are read and written using the library `weio`. 
+    The template file is read and its content can be changed like a dictionary.
+    Each item of `PARAMS` correspond to a set of parameters that will be replaced
+    in the template file to generate one input file.
+
+    For "FAST" input files, parameters can be changed recursively.
+    
+
+    INPUTS:
+      PARAMS: list of dictionaries. Each key of the dictionary should be a key present in the 
+              template file when read with `weio` (see: weio.read(main_file).keys() )
+
+               PARAMS[0]={'FAST|DT':0.1, 'EDFile|GBRatio':1, 'ServoFile|GenEff':0.8}
+
+      template_dir: if provided, this directory and its content will be copied to `output_dir` 
+                      before doing the parametric substitution
+
+      output_dir  : directory where files will be generated. 
     """
     def fileID(s):
         if s.find('|')<=0:
@@ -551,27 +625,39 @@ def templateReplaceGeneral(template_dir, PARAMS, workdir=None, main_file=None, n
         return os.path.splitext(os.path.basename(s))[0]
     def rebase(s,sid):
         split = os.path.splitext(os.path.basename(s))
-        return os.path.join(workdir,split[0]+sid+split[1])
+        return os.path.join(output_dir,split[0]+sid+split[1])
     def rebase_rel(s,sid):
         split = os.path.splitext(s)
-        return os.path.join(workdir,split[0]+sid+split[1])
-    # --- Saafety checks
-    if not os.path.exists(template_dir):
-        raise Exception('Template directory does not exist: '+template_dir)
+        return os.path.join(output_dir,split[0]+sid+split[1])
 
-    # Default value of workdir if not provided
-    if template_dir[-1]=='/'  or template_dir[-1]=='\\' :
-        template_dir=template_dir[0:-1]
-    if workdir is None:
-        workdir=template_dir+'_Parametric'
+    # --- Safety checks
+    if template_dir is None and output_dir is None:
+        raise Exception('Provide at least a template directory OR an output directory')
 
-    # Copying template folder to workdir
-    if os.path.exists(workdir) and RemoveAllowed:
-        shutil.rmtree(workdir, ignore_errors=False, onerror=handleRemoveReadonlyWin)
-    copyTree(template_dir, workdir)
+    if template_dir is not None:
+        if not os.path.exists(template_dir):
+            raise Exception('Template directory does not exist: '+template_dir)
 
-    # --- Fast main file use as "master"
-    main_file=os.path.join(workdir, os.path.basename(main_file))
+        # Default value of output_dir if not provided
+        if template_dir[-1]=='/'  or template_dir[-1]=='\\' :
+            template_dir=template_dir[0:-1]
+        if output_dir is None:
+            output_dir=template_dir+'_Parametric'
+
+    # --- Creating output_dir - Copying template folder to output_dir if necessary
+    if os.path.exists(output_dir) and RemoveAllowed:
+        shutil.rmtree(output_dir, ignore_errors=False, onerror=handleRemoveReadonlyWin)
+    if template_dir is not None:
+        copyTree(template_dir, output_dir)
+    else:
+        if not os.path.exists(output_dir):
+            os.makedirs(output_dir)
+
+    # --- Main file use as "master"
+    if template_dir is not None:
+        main_file=os.path.join(output_dir, os.path.basename(main_file))
+    else:
+        main_file=main_file
 
     # Params need to be a list
     if not isinstance(PARAMS,list):
@@ -595,7 +681,7 @@ def templateReplaceGeneral(template_dir, PARAMS, workdir=None, main_file=None, n
         # ---Copying main file and reading it
         #fst_full = rebase(main_file,strID)
         ext = os.path.splitext(main_file)[-1]
-        fst_full = os.path.join(workdir,strID+ext)
+        fst_full = os.path.join(output_dir,strID+ext)
         shutil.copyfile(main_file, fst_full )
         Files=dict()
         Files['ROOT']=weio.FASTInFile(fst_full)
@@ -606,9 +692,9 @@ def templateReplaceGeneral(template_dir, PARAMS, workdir=None, main_file=None, n
             if t=='ROOT':
                 continue
             org_filename      = Files['ROOT'][t].strip('"')
-            org_filename_full = os.path.join(workdir,org_filename)
+            org_filename_full = os.path.join(output_dir,org_filename)
             new_filename_full = rebase_rel(org_filename,'_'+strID)
-            new_filename      = os.path.relpath(new_filename_full,workdir)
+            new_filename      = os.path.relpath(new_filename_full,output_dir)
 #             print('org_filename',org_filename)
 #             print('org_filename',org_filename_full)
 #             print('New_filename',new_filename_full)
@@ -634,12 +720,14 @@ def templateReplaceGeneral(template_dir, PARAMS, workdir=None, main_file=None, n
             Files[t].write()
 
         files.append(fst_full)
+
     # --- Remove extra files at the end
-    os.remove(main_file)
+    if template_dir is not None:
+        os.remove(main_file)
 
     return files
 
-def templateReplace(template_dir, PARAMS, workdir=None, main_file=None, name_function=None, RemoveAllowed=False, RemoveRefSubFiles=False, oneSimPerDir=False):
+def templateReplace(PARAMS, template_dir, workdir=None, main_file=None, name_function=None, RemoveAllowed=False, RemoveRefSubFiles=False, oneSimPerDir=False):
     """ Replace parameters in a fast folder using a list of dictionaries where the keys are for instance:
         'FAST|DT', 'EDFile|GBRatio', 'ServoFile|GenEff'
     """
@@ -698,6 +786,10 @@ def templateReplace(template_dir, PARAMS, workdir=None, main_file=None, name_fun
             print(FstFiles)
             raise Exception('More than one fst file found in template folder, provide `main_file` or ensure there is only one `.fst` file') 
         main_file=FstFiles.pop()
+    # if the user provided a full path to the main file, we scrap the directory. TODO, should be cleaner
+    if len(os.path.dirname(main_file))>0:
+        main_file=os.path.basename(main_file)
+
 
 
     fastfiles=[]
@@ -786,7 +878,8 @@ def templateReplace(template_dir, PARAMS, workdir=None, main_file=None, name_fun
 # --------------------------------------------------------------------------------{
 def paramsSteadyAero(p=dict()):
     p['AeroFile|AFAeroMod']=1 # remove dynamic effects dynamic
-    #p['AeroFile|TwrPotent']=0 # remove tower shadow
+    p['AeroFile|WakeMod']=1 # remove dynamic inflow dynamic
+    p['AeroFile|TwrPotent']=0 # remove tower shadow
     return p
 
 def paramsNoGen(p=dict()):
@@ -992,7 +1085,14 @@ def averageDF(df,avgMethod='periods',avgParam=None,ColMap=None,ColKeep=None,ColS
         # NOTE: potentially we could average over each period and then average
         psi=df['Azimuth_[deg]'].values
         _,iBef = _zero_crossings(psi-psi[-10],direction='up')
-        tEnd = time[iBef[-1]]
+        if len(iBef)==0:
+            _,iBef = _zero_crossings(psi-180,direction='up')
+        if len(iBef)==0:
+            print('[WARN] Not able to find a zero crossing!')
+            tEnd = time[-1]
+            iBef=[0]
+        else:
+            tEnd = time[iBef[-1]]
         if avgParam is None:
             tStart=time[iBef[0]]
         else:
@@ -1000,8 +1100,11 @@ def averageDF(df,avgMethod='periods',avgParam=None,ColMap=None,ColKeep=None,ColS
             if len(iBef)-1<avgParam:
                 print('[WARN] Not enough periods found ({}) compared to number requested to average ({})!'.format(len(iBef)-1,avgParam))
                 avgParam=len(iBef)-1
-               
-            tStart=time[iBef[-1-avgParam]]
+            if avgParam==0:
+                tStart = time[0]
+                tEnd   = time[-1]
+            else:
+                tStart=time[iBef[-1-avgParam]]
     elif avgMethod.lower()=='periods_omega':
         # --- Using average omega to find periods
         if 'RotSpeed_[rpm]' not in df.columns:
@@ -1089,6 +1192,7 @@ def averagePostPro(outFiles,avgMethod='periods',avgParam=None,ColMap=None,ColKee
 # --------------------------------------------------------------------------------{
 def CPCT_LambdaPitch(refdir,main_fastfile,Lambda=None,Pitch=np.linspace(-10,40,5),WS=None,Omega=None, # operating conditions
           TMax=20,bStiff=True,bNoGen=True,bSteadyAero=True, # simulation options
+          ReRun=True, 
           fastExe=None,ShowOutputs=True,nCores=4): # execution options
     """ Computes CP and CT as function of tip speed ratio (lambda) and pitch.
     There are two main ways to define the inputs:
@@ -1097,6 +1201,10 @@ def CPCT_LambdaPitch(refdir,main_fastfile,Lambda=None,Pitch=np.linspace(-10,40,5
     """
 
     WS_default=5 # If user does not provide a wind speed vector, wind speed used
+
+    # if the user provided a full path to the main file, we scrap the directory. TODO, should be cleaner
+    if len(os.path.dirname(main_fastfile))>0:
+        main_fastfile=os.path.basename(main_fastfile)
 
     # --- Reading main fast file to get rotor radius 
     fst = weio.FASTInFile(os.path.join(refdir,main_fastfile))
@@ -1143,11 +1251,12 @@ def CPCT_LambdaPitch(refdir,main_fastfile,Lambda=None,Pitch=np.linspace(-10,40,5
     # --- Generating all files in a workdir
     workdir = refdir.strip('/').strip('\\')+'_CPLambdaPitch'
     print('>>> Generating inputs files in {}'.format(workdir))
-    fastFiles=templateReplace(refdir,PARAMS,workdir=workdir,name_function=naming,RemoveRefSubFiles=True,RemoveAllowed=True,main_file=main_fastfile)
+    RemoveAllowed=ReRun # If the user want to rerun, we can remove, otherwise we keep existing simulations
+    fastFiles=templateReplace(PARAMS, refdir, workdir=workdir,name_function=naming,RemoveRefSubFiles=True,RemoveAllowed=RemoveAllowed,main_file=main_fastfile)
 
     # --- Running fast simulations
     print('>>> Running {} simulations...'.format(len(fastFiles)))
-    run_fastfiles(fastFiles, ShowOutputs=ShowOutputs, fastExe=fastExe, nCores=nCores)
+    run_fastfiles(fastFiles, ShowOutputs=ShowOutputs, fastExe=fastExe, nCores=nCores, ReRun=ReRun)
 
     # --- Postpro - Computing averages at the end of the simluation
     print('>>> Postprocessing...')
@@ -1214,5 +1323,5 @@ if __name__=='__main__':
     PARAMS['ServoFile|VS_Rgn2K']    = 0.00038245
     PARAMS['ServoFile|GenEff']      = 0.95
     PARAMS['InflowFile|HWindSpeed'] = 8
-    templateReplace(ref_dir,PARAMS,name_function=naming,RemoveRefSubFiles=True)
+    templateReplace(PARAMS,ref_dir,name_function=naming,RemoveRefSubFiles=True)
